% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-sage.R
\name{display_sage}
\alias{display_sage}
\title{Display tour path with a sage scatterplot}
\usage{
display_sage(
  axes = "center",
  half_range = NULL,
  col = "black",
  pch = 20,
  gam = 1,
  R = NULL,
  ...
)
}
\arguments{
\item{axes}{position of the axes: center, bottomleft or off}

\item{half_range}{half range to use when calculating limits of projected.
If not set, defaults to maximum distance from origin to each row of data.}

\item{col}{color to be plotted.  Defaults to "black"}

\item{pch}{marker for points. Defaults to 20.}

\item{gam}{scaling of the effective dimensionality for rescaling. Defaults to 1.}

\item{R}{scale for the radial transformation.
If not set, defaults to maximum distance from origin to each row of data.}

\item{...}{other arguments passed on to \code{\link{animate}} and
\code{\link{display_sage}}}
}
\description{
Animate a 2D tour path with a sage scatterplot that
uses a radial transformation on the projected points to re-allocate
the volume projected across the 2D plane.
}
\examples{
# Generate samples on a 3d and 5d hollow sphere using the geozoo package
sphere3 <- geozoo::sphere.hollow(3)$points
sphere5 <- geozoo::sphere.hollow(5)$points

# Columns need to be named before launching the tour
colnames(sphere3) <- c("x1", "x2", "x3")
colnames(sphere5) <- c("x1", "x2", "x3", "x4", "x5")

}
