% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sketches}
\alias{sketches}
\title{Sketches from Quickdraw
6000 random images from six categories in the QuickDraw}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6000 rows and 786 columns.
}
\source{
https://github.com/huizezhang-sherry/quickdraw, https://quickdraw.withgoogle.com/data
}
\usage{
sketches
}
\description{
The columns correspond to grayscale pixel values (V1-V784) of a 28 pixel
by 28 pixel image, alongside the word (category) and the identifier for
the image.
}
\keyword{datasets}
